+++ CHANGELOG.rst
@@ -53,7 +53,8 @@ Release notes
   new `data` dict.
   https://github.com/aboutcode-org/dejacode/issues/202
 
-- Add `vulnerabilities_risk_threshold` field to the DataspaceConfiguration model.
+- Add the `vulnerabilities_risk_threshold` field to the Product and
+  DataspaceConfiguration models.
   This threshold helps prioritize and control the level of attention to vulnerabilities.
   https://github.com/aboutcode-org/dejacode/issues/97